add frontend and backend validation for size/format

+++ apps/meteor/app/custom-sounds/server/methods/uploadCustomSound.ts
@@ -1,9 +1,12 @@
 import { api } from '@rocket.chat/core-services';
 import type { RequiredField } from '@rocket.chat/core-typings';
 import type { ServerMethods } from '@rocket.chat/ddp-client';
+import { fromBuffer } from 'file-type';
 import { Meteor } from 'meteor/meteor';
 
 import type { ICustomSoundData } from './insertOrUpdateSound';
+import { MAX_CUSTOM_SOUND_SIZE_BYTES } from '../../../../lib/constants';
+import { MIME } from '../../../../server/ufs/ufs-mime';
 import { hasPermissionAsync } from '../../../authorization/server/functions/hasPermission';
 import { RocketChatFile } from '../../../file/server';
 import { RocketChatFileCustomSoundsInstance } from '../startup/custom-sounds';
@@ -22,6 +25,14 @@ Meteor.methods<ServerMethods>({
 		}
 
 		const file = Buffer.from(binaryContent, 'binary');
+		if (file.length > MAX_CUSTOM_SOUND_SIZE_BYTES) {
+			throw new Meteor.Error('file-too-large', 'Max size is 5MB');
+		}
+
+		const mimeType = await fromBuffer(file);
+		if (!mimeType || mimeType.mime !== MIME.mp3 || mimeType.ext !== 'mp3') {
+			throw new Meteor.Error('invalid-file-type', 'Only MP3 files are allowed');
+		}
 
 		const rs = RocketChatFile.bufferToStream(file);
 		await RocketChatFileCustomSoundsInstance.deleteFile(`${soundData._id}.${soundData.extension}`);
